requirements.txt
alpaca-py>=0.43.0
pandas>=2.0.0
numpy>=2.0.0
pyarrow>=14.0.0
ta>=0.11.0
anthropic>=0.98.0
ccxt>=4.0.0
feedparser>=6.0.0
schedule>=1.2.0
aiohttp>=3.9.0
requests>=2.31.0
websocket-client>=1.6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
pytest>=9.0.0
